add singlton effect list and refactoring button processing
