+++ src/button/button_handler.h
@@ -1,0 +1,2 @@
+void setup_buttons();
+void tick_buttons();
